py/pyproject.toml
[project]
name = "xrr"
version = "0.1.0"
description = "Multi-channel interaction recorder/replayer"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/xrr"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
